+++ bia-ingest/bia_ingest/ingest/file_reference.py
@@ -2,17 +2,20 @@ import logging
 from pydantic import ValidationError
 from typing import List, Dict, Optional
 
+from .biostudies.submission_parsing_utils import (
+    attributes_to_dict,
+)
+
 from ..bia_object_creation_utils import dict_to_uuid, filter_model_dictionary
-from .utils import (
+from .biostudies.submission_parsing_utils import (
     find_datasets_with_file_lists,
 )
-from .biostudies import (
+from .biostudies.api import (
     Submission,
-    attributes_to_dict,
-    flist_from_flist_fname,
     file_uri,
+    flist_from_flist_fname,
 )
-from . import biostudies  # To make reference to biostudies.File explicit
+from .biostudies import api  # To make reference to biostudies.File explicit
 from bia_shared_datamodels import bia_data_model
 from ..persistence_strategy import PersistenceStrategy
 
@@ -85,7 +88,7 @@ def get_file_reference_for_submission_dataset(
         bia_data_model.ExperimentalImagingDataset
         | bia_data_model.ImageAnnotationDataset
     ),
-    files_in_file_list: List[biostudies.File],
+    files_in_file_list: List[api.File],
     result_summary: dict,
 ) -> List[bia_data_model.FileReference]:
     """
