+++ bia-ingest/bia_ingest/ingest/specimen.py
@@ -3,17 +3,20 @@ import logging
 from typing import List, Any, Dict, Optional
 from bia_shared_datamodels import bia_data_model
 
-from ..bia_object_creation_utils import dict_to_uuid, filter_model_dictionary
+from ..bia_object_creation_utils import (
+    dict_to_uuid,
+    dicts_to_api_models,
+    filter_model_dictionary,
+)
 
 from ..cli_logging import log_failed_model_creation, log_model_creation_count
 from ..persistence_strategy import PersistenceStrategy
 from pydantic import ValidationError
 from .utils import (
-    dicts_to_api_models,
     get_generic_section_as_list,
     object_value_pair_to_dict,
 )
-from .biostudies import (
+from .biostudies.api import (
     Submission,
 )
 from . import (
