+++ bia-ingest/bia_ingest/ingest/study.py
@@ -3,17 +3,20 @@ from pydantic import ValidationError
 import re
 from typing import List, Any, Dict, Optional
 
+from .biostudies.submission_parsing_utils import (
+    attributes_to_dict,
+    find_sections_recursive,
+    mattributes_to_dict,
+)
+
 from ..bia_object_creation_utils import dict_to_uuid
 
 from ..cli_logging import log_failed_model_creation
 from .utils import (
     get_generic_section_as_dict,
-    mattributes_to_dict,
-    find_sections_recursive,
 )
-from .biostudies import (
+from .biostudies.api import (
     Submission,
-    attributes_to_dict,
 )
 from ..persistence_strategy import PersistenceStrategy
 from bia_shared_datamodels import bia_data_model, semantic_models
