+++ bia-ingest/test/conftest.py
@@ -3,7 +3,7 @@ from typing import Dict
 from pathlib import Path
 import json
 import pytest
-from bia_ingest.ingest.biostudies import Submission, SubmissionTable, requests
+from bia_ingest.ingest.biostudies.api import Submission, SubmissionTable, requests
 from .utils import accession_id
 from bia_ingest.cli_logging import IngestionResult
